correct mistake

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
-    public double calculateBmi(double height, double weight) {
+    public int calculateBmi(double height, int weight) {
         double bmi = weight / Math.pow(height, 2);
         return (int) bmi;
     }
